requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.10
Django==5.1.7
django-crispy-forms==2.3
django-environ==0.12.0
psycopg2==2.9.10
sqlparse==0.5.3
wget==3.2
